requirements.txt
python-telegram-bot==21.0
peewee==3.17.0
python-dotenv==1.0.0
